Refactor user interaction operators to User.java
Moved user interaction related operators (e.g., clipboard, command execution, user input dialogs, wizard, and value entry) from System.java to a new User.java file. This improves separation of concerns and keeps System.java focused on system-level operations.
