+++ gama.ui.shared/src/gama/ui/shared/parameters/MonitorDisplayer.java
@@ -13,8 +13,8 @@ package gama.ui.shared.parameters;
 import static gama.core.runtime.GAMA.getRuntimeScope;
 import static gama.core.runtime.GAMA.reportError;
 import static gama.core.util.GamaListFactory.wrap;
-import static gama.gaml.operators.System.enterValue;
-import static gama.gaml.operators.System.userInputDialog;
+import static gama.gaml.operators.User.enterValue;
+import static gama.gaml.operators.User.userInputDialog;
 import static gama.gaml.types.Types.NO_TYPE;
 import static gama.ui.shared.menus.GamaMenu.action;
 import static gama.ui.shared.menus.GamaMenu.separate;
